imported missing functions
