+++ main.py
@@ -8,9 +8,9 @@ from discord.ext import tasks
 from dotenv import load_dotenv
 
 import helpers.globals as globals
-from helpers.notifications import load_filter_data
+from helpers.notifications import load_filter_data, read_json_data, build_filter_message
 from helpers.environment import prepare_environment
-from helpers.quests import fetch_today_data
+from helpers.quests import fetch_today_data, find_quest, write_filter_data
 
 # Validates arguments passed to check what env was requested
 if (len(sys.argv) != 2):
@@ -27,91 +27,6 @@ muted_role = ""
 
 comandoQuestsTitle = "__COMANDOS IMPLEMENTADOS:__"
 comandoQuestsBody = "> !questleiria/questmarinha POKÉSTOP/QUEST/RECOMPENSA\nDevolve uma lista de resultados onde a pokéstop, quest ou recompensa correspondam ao texto inserido\n`(ex:!questmarinha startdust | !questleiria tribunal)`"
-
-def build_quest_message(data):
-    return "[" + data['name'] + "](" + build_quest_location_url(data["latitude"], data["longitude"]) + ")"
-
-def build_quest_location_url(latitude, longitude):
-    coordinatesUrl = "https://www.google.com/maps/search/?api=1&query=" + str(latitude) + "," + str(longitude)
-
-    return coordinatesUrl
-
-def write_filter_data(receivedData, add=True):
-    if len(receivedData) != 2:
-        return False
-
-    originalDiscordChannelName = {"raros": ["pokemon", "pokemonmarinha"], "uteis": ["pokemonuteis", "pokemonuteismarinha"]}
-    filterName = {"raros": "spawns-raros", "uteis": "spawns-uteis"}
-
-    with open(globals.FILTER_FILE) as raw_data:
-        jsonPokemonData = json.load(raw_data)
-
-    try:
-        pokemon = receivedData[0].title()
-        affectedChannel = originalDiscordChannelName[receivedData[1]]
-    except:
-        return "Não reconheço esse comando. Aqui tens uma lista para te ajudar." + "\n\n__**COMANDOS IMPLEMENTADOS:**__\n\n> !add POKEMON CANAL\nPara se adicionar um Pokémon a um canal específico `(ex:!add Poliwag uteis)`\n> !remove POKEMON CANAL\nPara que se remova um Pokemon de um canal específico `(ex:!remove Poliwag raros)`\n> !reload\nApós se alterar a lista, podem reiniciar as notificações com as alterações usando `(ex:!reload)`"
-
-    for name in affectedChannel:
-        if add and pokemon not in jsonPokemonData['monsters']['filters'][name]['monsters']:
-            jsonPokemonData['monsters']['filters'][name]['monsters'].append(pokemon)
-        elif not add and pokemon in jsonPokemonData['monsters']['filters'][name]['monsters']:
-            jsonPokemonData['monsters']['filters'][name]['monsters'].remove(pokemon)
-
-    os.remove(globals.FILTER_FILE)
-
-    with open(globals.FILTER_FILE, 'w') as file:
-        json.dump(jsonPokemonData, file, indent=4)
-
-    return pokemon + (" adicionado a " if add else " removido de ") + filterName[receivedData[1]]
-
-def find_quest(receivedData, local):
-    if receivedData and (receivedData == "!questleiria" or receivedData == "!questmarinha"):
-        return False
-
-    with open(globals.QUESTS_FILE) as raw_data:
-        quests = json.load(raw_data)
-    quests = sorted(quests, key=lambda k: k['quest_task'], reverse=True)
-
-    allQuestData = []
-    allQuestDataMarinha = []
-
-    for quest in quests:
-        if quest['quest_reward_type'] == 'Pokemon':
-            reward = quest['pokemon_name']
-        elif quest['quest_reward_type'] == 'Item':
-            reward = str(quest['item_amount']) + " " + quest['item_type']
-        elif quest['quest_reward_type'] == 'Stardust':
-            reward =  str(quest['item_amount']) + " " + quest['quest_reward_type']
-        else:
-            reward =  str(quest['item_amount']) + " " + quest['pokemon_name'] + " " + quest['item_type']
-
-        try:
-            if receivedData.lower() in reward.lower() or receivedData.lower() in quest['quest_task'].lower() or receivedData.lower() in quest["name"].lower():
-                if "-8.9" not in str(quest['longitude']):
-                    allQuestData.append({
-                        "name": "[Leiria] " + quest["name"],
-                        "map": build_quest_location_url(quest["latitude"], quest["longitude"]),
-                        "quest": quest['quest_task'] + " - " + reward,
-                        "image": quest["url"]
-                    })
-                else:
-                    allQuestDataMarinha.append({
-                        "name": "[Marinha Grande] " + quest["name"],
-                        "map": build_quest_location_url(quest["latitude"], quest["longitude"]),
-                        "quest": quest['quest_task'] + " - " + reward,
-                        "image": quest["url"]
-                    })
-        except Exception as e:
-            print(e)
-            return "Essa procura não me parece ser válida."
-
-    if local == 1:
-        return allQuestData
-
-    return allQuestDataMarinha
-
-
 
 def get_version():
     #with open('savedVersion.txt', 'r') as data:
